requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.11.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.1.0
Authlib==1.2.1
Babel==2.12.1
beautifulsoup4==4.12.2
black==23.3.0
blinker==1.6.2
cachelib==0.9.0
certifi==2023.5.7
cffi==1.16.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==41.0.3
cssselect==1.2.0
defusedxml==0.7.1
dnspython==2.3.0
dominate==2.8.0
elastic-transport==8.4.0
elasticsearch==8.8.0
email-validator==2.0.0.post2
favicon==0.7.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.15.4
Flask==2.3.2
Flask-Babel==2.0.0
Flask-Bootstrap==3.3.7.1
Flask-Caching==2.3.0
Flask-Cors==4.0.0
flask-htmx==0.3.1
Flask-HTTPAuth==4.8.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-Moment==1.0.5
Flask-PageDown==0.4.0
Flask-SQLAlchemy==3.0.4
Flask-WTF==1.1.1
frozenlist==1.4.1
greenlet==3.0.3
gunicorn==21.2.0
httpie==3.2.2
idna==3.4
importlib-metadata==6.6.0
isort==5.13.2
itsdangerous==2.1.2
jieba3k==0.35.1
Jinja2==3.1.2
joblib==1.4.2
langdetect==1.0.9
lxml==4.9.3
Mako==1.2.4
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
newspaper3k==0.2.8
nltk==3.8.1
openai==0.28.0
packaging==23.1
pathspec==0.12.1
Pillow==10.0.0
platformdirs==4.2.2
pycparser==2.21
Pygments==2.16.1
PyJWT==2.7.0
pymdown-extensions==10.9
PyMySQL==1.1.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
redis==5.0.7
regex==2024.5.15
requests==2.31.0
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==13.5.2
setuptools==71.0.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.16
tinysegmenter==0.3
tldextract==5.1.2
tomli==2.0.1
tqdm==4.66.1
typing_extensions==4.6.3
urllib3==1.26.16
validators==0.21.2
visitor==0.1.3
Werkzeug==2.3.6
WTForms==3.0.1
yarl==1.9.4
zipp==3.15.0
psycopg2-binary==2.9.9
scikit-learn==1.6.1
